feat: windows builds

+++ src/net/net_internal.h
@@ -13,22 +13,22 @@ struct ni_wi_payload {
     int8_t count;
 };
 
-void net_read_buf(void *dest, size_t n);
+void net_read_buf(void *dest, int n);
 byte net_read_byte(void);
 short net_read_short(void);
 int net_read_int(void);
-long net_read_long(void);
+int64_t net_read_long(void);
 float net_read_float(void);
 double net_read_double(void);
 bool net_read_bool(void);
 string8 net_read_string8(void);
 string16 net_read_string16(void);
 
-void net_write_buf(const void *buf, size_t n);
+void net_write_buf(const void *buf, int n);
 void net_write_byte(ubyte v);
 void net_write_short(short v);
 void net_write_int(int v);
-void net_write_long(long v);
+void net_write_long(int64_t v);
 void net_write_float(float v);
 void net_write_double(double v);
 void net_write_bool(bool v);
